Change constructor interface
The constructor now gets three obligatory positional arguments: URL,
USERNAME, and PASSWORD. An optional fourth argument may be used to
pass options to the underlying REST::Client object. This simplified
the credential passing for the user.

The constructor argument types are checked for consistency.

The "Accept: application/json" header is always passed. I removed the
option to have it optional because that was an overdesign.

Signed-off-by: Gustavo L. de M. Chaves <gnustavo@cpan.org>